Noizu.ElixirCore.CallingContext (noizu_core v1.0.26)

CallingContext

Context object used to track a caller's state and permissions along with unique request identifier for tracking requests as they travel through the layers in your application. Useful for log collation, permission checks, access auditing.

Application Config Options and Defaults

Include Default IO.inspect formatter defimpl for CallingContext config :noizu_core, inspect_calling_context: true

Request ID Extraction Strategy. {m,f} or function2 that accepts (conn, default) and pulls request id (CallingContext.token) from Plug.Conn. config :noizu_core, token_strategy: {Noizu.ElixirCore.CallingContext, :extract_token} - Request Reason Extraction Strategy. {m,f} or function2 that accepts (conn, default) and pulls request reason (CallingContext.reason) from Plug.Conn.config :noizu_core, token_strategy: {Noizu.ElixirCore.CallingContext, :extract_reason}

Request Caller Extraction Strategy. {m,f} or function2 that accepts (conn, default) and pulls request Caller (CallingContext.caller) from Plug.Conn config :noizu_core, get_plug_caller: {Noizu.ElixirCore.CallingContext, :extract_caller} - Request Caller's Auth Map Extraction Strategy. {m,f} or function2 that accepts (conn, default) and returns effective permission map ContextCaller.auth.config :noizu_core, acl_strategy: {Noizu.ElixirCore.CallingContext, :default_auth}

Default Blank/Empty Request Reason config :noizu_core, default_request_reason: :none

Default Restricted User Reference or Object config :noizu_core, default_internal_user: {:ref, Noizu.ElixirCore.CallerEntity, :restricted}

Default Internal User Reference or Object config :noizu_core, default_internal_user: {:ref, Noizu.ElixirCore.CallerEntity, :internal}

Default System User Reference or Object config :noizu_core, default_system_user: {:ref, Noizu.ElixirCore.CallerEntity, :system}

Default Admin User Reference or Object config :noizu_core, default_admin_user: {:ref, Noizu.ElixirCore.CallerEntity, :admin}

Default Restricted User Auth Map config :noizu_core, default_internal_auth: %{permissions: %{restricted: true}}

Default Internal User Auth Map config :noizu_core, default_internal_auth: %{permissions: %{internal: true}}

Default Ssy User Auth Map config :noizu_core, default_system_auth: %{permissions: %{system: true, internal: true}}

Default Admin User Auth Map config :noizu_core, default_admin_auth: %{permissions: %{admin: true, system: true, internal: true}}

Types

t()

@type t() :: %Noizu.ElixirCore.CallingContext{
  auth: Any,
  caller: tuple(),
  options: Map.t(),
  outer_context: CalingContext.t(),
  reason: String.t(),
  time: DateTime.t() | nil,
  token: String.t(),
  vsn: float()
}

Functions

admin()

Create new calling context with default admin user caller and permissions.

admin(this)

admin(conn, options)

default_reason(conn, default)

Default logic for grabbing conn's request reason. Using either config noizu_core, extract_reason: {m,f} or included extract_token method.

default_token(conn, default)

Default logic for grabbing conn's request id. Using either config noizu_core, token_strategy: {m,f} or included extract_token method.

extract_caller(conn, default)

Extract caller permission level from Plug.Conn. (Currently only depends on default value)

extract_reason(conn, default)

Extract Request Reason from Plug.Conn. Useful for logging purposes/exception handling messaging/tracing on backend.

extract_token(conn, default)

Extract Request Id Token from Plug.Conn or Generate UUID if non found. Attached to CallingContext to include in log output for log collation.

get_auth(caller, options)

Get caller effective authentication map.

get_caller(conn, default \\ :unauthenticated)

Get caller from Plug.Conn

get_reason(conn, default \\ nil)

Get call reason from Plug.Conn

get_token(conn, default \\ :generate)

Get call request id from Plug.Conn

internal()

Create new calling context with default internal user caller and permissions.

internal(this)

internal(conn, options)

meta_strip(context)

Strip CallingContext meta data from Logger.

meta_update(context)

Update Logger with CallingContext meta data.

metadata(context)

Extract CallingContext meta data for Logger

new(caller, auth, options)

Generate new Context.Caller by passing in Caller, Auth and options.

new_conn(conn, options \\ %{logger_init: true})

Prepare Context object with for given conn. Sets Process meta data by default.

new_conn(caller, auth, conn, options \\ %{logger_init: true})

Prepare Context object with given caller/auth and conn request details. Sets Process meta data by default.

restricted()

Create new calling context with default restricted user caller and permissions.

restricted(this)

restricted(conn, options)

system()

Create new calling context with default system user caller and permissions.

system(this)

system(conn, options)