Noizu.FastGlobal.Database (noizu_core v1.0.26)

The Noizu.FastGlobal.Database module defines the schema for the FastGlobal Database.

This module uses the Amnesia library, a simple and powerful library for working with Mnesia, the Erlang's built-in distributed real-time database.

Tables

The database consists of the following tables:

Settings: stores the settings for the FastGlobal database.

Code Review

The code is simple and straightforward, following the conventions of the Amnesia library. The schema definition is clear, and the use of the deftable macro makes it easy to understand the structure of the database tables. The Settings table is well-defined, with clearly named fields. The use of the type: :set option indicates that each entry in the table should be unique. However, it should be noted that no indexes are defined for the table, which may impact performance if the table grows large and queries are performed on non-key fields. The @type specification for the Settings struct is a good practice, as it provides type checking and documentation benefits.

Functions

__using__(opts)
(macro)

Alias all the table names in the current scope and require what's needed.

create!(copying \\ [])

@spec create!(Amnesia.Table.c()) :: [Amnesia.Table.o()]

Create the database, it calls .create on every defined table, raises in case of error.

create(copying \\ [])

@spec create(Amnesia.Table.c()) :: [Amnesia.Table.o()]

Create the database, it calls .create on every defined table.

destroy!()

@spec destroy!() :: [Amnesia.Table.o()]

Destroy the database, it calls .destroy on every defined table, raises in case of error.

destroy()

@spec destroy() :: [Amnesia.Table.o()]

Destroy the database, it calls .destroy on every defined table.

metadata()

@spec metadata() :: Amnesia.Metadata.t()

tables()

@spec tables() :: [atom()]

List of atoms of the defined tables in the database.

wait(timeout \\ :infinity)

@spec wait(integer() | :infinity) :: :ok | {:timeout, [atom()]} | {:error, atom()}

Wait for the database to be loaded.